feat: reformat code architecture to add indexer in the same repo
