+++ keeper/resources/satoru_DepositHandler.contract_class.json
