+++ keeper/resources/satoru_OrderHandler.compiled_contract_class.json
